+++ src/components/icons.tsx
@@ -92,3 +92,11 @@ export function IconBrand(props: IconProps) {
     </Icon>
   );
 }
+
+export function IconCheck(props: IconProps) {
+  return (
+    <Icon {...props}>
+      <polyline points="20 6 9 17 4 12" />
+    </Icon>
+  );
+}